Refactor tests to use PSR-4 naming and namespaces
Modernise test file structure to follow PSR-4 conventions with proper
namespacing under Automattic\EditFlow\Tests\Integration.

Changes:
- Rename test files from test-edit-flow-*.php to *Test.php pattern
- Rename test classes to match file names (e.g., CalendarTest)
- Add namespace Automattic\EditFlow\Tests\Integration to all tests
- Rename testcase-edit-flow-ajax.php to AjaxTestCase.php
- Update phpunit.xml.dist to use suffix="Test.php"
- Remove obsolete .phpunit.result.cache from .gitignore
