+++ phpunit.xml.dist
@@ -25,7 +25,7 @@
 			<directory suffix="Test.php">tests/Unit</directory>
 		</testsuite>
 		<testsuite name="integration">
-			<directory prefix="test-" suffix=".php">tests/Integration</directory>
+			<directory suffix="Test.php">tests/Integration</directory>
 		</testsuite>
 	</testsuites>
