+++ tests/Integration/DashboardNotepadTest.php
@@ -1,7 +1,19 @@
 <?php
+/**
+ * Dashboard Notepad integration tests.
+ *
+ * @package Automattic\EditFlow\Tests\Integration
+ */
 
-class WP_Test_Edit_Flow_Dashboard_Note extends WP_UnitTestCase {
-	
+declare( strict_types=1 );
+
+namespace Automattic\EditFlow\Tests\Integration;
+
+use EF_Dashboard_Notepad_Widget;
+use WP_UnitTestCase;
+
+class DashboardNotepadTest extends WP_UnitTestCase {
+
 	function test_register_dashboard_note_post_type() {
 		//As part of the Edit Flow initialziation process
 		//EF_Dashboard_Notepad_Widget should have already
@@ -9,7 +21,7 @@ class WP_Test_Edit_Flow_Dashboard_Note extends WP_UnitTestCase {
 		$pobj = get_post_type_object( 'dashboard-note' );
 		$this->assertEquals( 'dashboard-note', $pobj->name );
 
-		//Testing EF_Dashboard_Notepad_Widget::init explicitly 
+		//Testing EF_Dashboard_Notepad_Widget::init explicitly
 		_unregister_post_type( 'dashboard-note' );
 
 		$EditFlowDashboardNote = new EF_Dashboard_Notepad_Widget();
