+++ tests/Integration/StarterTest.php
@@ -1,28 +1,32 @@
 <?php
 /**
- * Tests to test that that testing framework is testing tests. Meta, huh?
- * @package wordpress-plugins-tests
- * @author mbijon
+ * Starter tests to verify testing framework is working.
+ *
+ * @package Automattic\EditFlow\Tests\Integration
  */
-class WP_Test_Edit_Flow_Starter_Tests extends WP_UnitTestCase {
-	
+
+declare( strict_types=1 );
+
+namespace Automattic\EditFlow\Tests\Integration;
+
+use WP_UnitTestCase;
+
+class StarterTest extends WP_UnitTestCase {
+
 	/**
 	 * Run a simple test to ensure that the tests are running
 	 */
 	function test_editflow_exists() {
-		 
 		$this->assertTrue( class_exists( 'edit_flow' ) );
-		 
 	}
-	
+
 	/**
 	 * Verify a minimum version of WordPress is installed
 	 */
 	function test_wp_version() {
-		
 		$minimum_version = '3.4.0';
 		$running_version = get_bloginfo( 'version' );
-		
+
 		//trunk is always "master" in github terms, but WordPress has a specific way of describing it
 		//grab the exact version number to verify that we're on trunk
 		if ( $running_version == 'master' || $running_version == 'trunk' ) {
@@ -30,23 +34,19 @@ class WP_Test_Edit_Flow_Starter_Tests extends WP_UnitTestCase {
 			preg_match( '#\$wp_version = \'([^\']+)\';#', $file, $matches );
 			$running_version = $matches[1];
 		}
-		
+
 		$this->assertTrue( version_compare( $running_version, $minimum_version, '>=' ) );
-	
 	}
-	
+
 	/**
 	 * Test modules loading
 	 */
 	function test_editflow_register_module() {
-		
-		$EditFlow = EditFlow();
-		
+		$EditFlow = \EditFlow();
+
 		$module_real = strtolower( 'calendar' );
 		$module_args = array ( 'title' => $module_real );
 		$module_return = $EditFlow->register_module( $module_real, $module_args );
 		$this->assertTrue( $module_real == $module_return->name );
-		
 	}
-	
 }
